+++ apps/doop/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @cloudoperators/juno-app-doop
 
+## 2.4.3
+
+### Patch Changes
+
+- 86e0be2: Add unique id to items for stable ReactMarkdown rendering and fix inconsistent instance content
+- Updated dependencies [b693d90]
+  - @cloudoperators/juno-url-state-provider@3.0.3
+
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies [5488568]
+- Updated dependencies [6ef7f8d]
+  - @cloudoperators/juno-ui-components@4.0.1
+  - @cloudoperators/juno-messages-provider@0.2.8
+
+## 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [7317372]
+  - @cloudoperators/juno-messages-provider@0.2.7
+
 ## 2.4.0
 
 ### Minor Changes